Temperature Converter Website Task-3 Web Development
The Temperature Converter website is a simple yet effective tool designed to convert temperatures between Fahrenheit, Celsius, and Kelvin units. The user-friendly interface allows you to input a temperature value and select the desired conversion type from a dropdown menu. The website features a clean and visually appealing design, with a responsive layout for seamless use on various devices.

Key Features:
	Supports conversion between Fahrenheit, Celsius, and Kelvin.
	Intuitive input field and dropdown menu for easy selection.
	Stylish and responsive design for an enhanced user experience.
	Real-time temperature conversion results displayed dynamically.
	Hover effects and transitions for a modern and engaging interface.
The website is built using HTML for the structure, CSS for styling, and JavaScript for the temperature conversion logic. It includes image elements to enhance the visual appeal and engages users with a smooth transition effect when hovering over the temperature conversion container. This Temperature Converter serves as a practical tool for quick and accurate temperature conversions, suitable for a wide range of users and applications.